+++ carsus/io/output/__init__.py
@@ -1,1 +1,1 @@
-from tardis_op import AtomData
+from tardis_ import AtomData
